requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.5
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==25.2
django-braces==1.17.0
django-cors-headers==4.7.0
django-csp==4.0
django-extensions==4.1
django-filter==25.1
django-multiselectfield==1.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dlib==20.0.0
drf-spectacular==0.28.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fonttools==4.59.2
frozenlist==1.7.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lxml==6.0.0
matplotlib==3.10.6
MouseInfo==0.1.3
multidict==6.6.4
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pdfkit==1.0.0
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-magic-bin==0.4.14
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
rpds-py==0.27.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
twilio==9.8.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.9.0
xlsxwriter==3.2.5
yarl==1.20.1
zope.interface==7.2
